+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.1.8 (2019-08-06)
 
 * Added `stick` property.
-* Changed `NO`, `TOP_LEFT` and `TOP_RIGHT` to `no`, `topLeft`, `topRight`. 
+* Changed `NO`, `TOP_LEFT` and `TOP_RIGHT` to `no`, `leftTop`, `rightTop` in BubbleNip. 
 * Refactoring.
 
 ## 1.1.7 (2019-08-05)